added mapping to boolean, float, long and double genes for strings with stringSpecialization with the corresponding type

+++ client-java/instrumentation/src/main/java/org/evomaster/client/java/instrumentation/coverage/methodreplacement/classes/DoubleClassReplacement.java
@@ -26,7 +26,7 @@ public class DoubleClassReplacement implements MethodReplacementClass {
 
         if (TaintInputName.isTaintInput(input)) {
             ExecutionTracer.addStringSpecialization(input,
-                    new StringSpecializationInfo(StringSpecialization.INTEGER, null));
+                    new StringSpecializationInfo(StringSpecialization.DOUBLE, null));
         }
 
         if (idTemplate == null) {
